requirements.txt
google-adk==0.2.0
fastapi>=0.115.0
pydantic>=2.5.2
requests>=2.31.0
uvicorn>=0.24.0
openai>=1.0.0
litellm